ver em casa yessir

+++ include/sequential/queue.h
@@ -14,6 +14,6 @@ bool q_pop(Queue *q, unsigned short *out_value);
 bool q_front(const Queue *q, unsigned short *out_value);
 
 bool q_is_empty(const Queue *q);
-unsigned short q_size(const Queue *q);
+unsigned long long q_size(const Queue *q);
 
 #endif
